Unbind bounds checker before doing a manual zoom
Without this change the manual zoom was picked up and the bounds checker
then went and re-drew the markers on the map for the current zoom level.
This means that opening a popup manually didn't work properly. By
unbinding before we zoom to a provider, then re-binding after we avoid
the re-draw.