Disable pipeline during browser based tests
For some reason using the pre-compiled assets during these tests doesn't
work properly. It seems to be an incompatibility between pipeline and
LiveServerTestCase, something to do with the way static files are
handled in those tests, though I'm not sure exactly what.

This makes the browser tests run with pipeline disabled (which means the
sass gets compiled before each test).