Add a row_href attribute to the ReviewTable class.
Fixes #1083 by giving the table the method to get a link for the current row. To make this work I had to add another parameter to the __init__ to take the current cobrand, so that we knew what to use in making the link. In doing so, I also noticed that the fallback href on the content hardcoded the choices cobrand, so I fixed this too