Fix the service select box not being properly disabled when the org type changed.
WHat was happening was that the selector was wrong in the org select box's change event handler, so it was never disabling the selection box after the org type changed. I also realised that now that we have two org types with services, the previous code which tried to disable the org type selection once a service selection was made no longer made any sense. I've removed that bit, so you can change the org type at will once you've selected a service, but if you pick 'All' or 'GP' it wipes the service selection, since that's invalid. I also changed the way selected options are shown, so that they show immediately, rather than on reload