Trigger update instead of change on org_type filter when resetting it after picking an org directly.
Triggering change causes all the map pins to be reloaded, which creates a race between that and reloading all the pins because we selected an organisation. During that race, quite often the newly selected org would have been located and its' popup open, only for the filter-triggered reload to come in an wipe it out, hiding the popup you wanted to see.

By triggering update instead, we can just update the displayed value without having to reload the pins as well, which is unnecessary anyway since they're just about to be reloaded anyway.Closes #1210