Fix bug which wiped out users on CCGs and Trusts.
You can't assign an empty list to a related object manager like steve - Django interprets this as you clearing the objects and dutifully wipes them off the database. Even making a deepcopy of the object and working on that doesn't stop it. Rather than faff more I decided the safest way was to specify fields to the serializer, rather than edit the content of the fields