Combine review date fields into one select box.
Having two select boxes means that we have to allow some inputs
that are then refused by the validation for either being twoo
old or in the future, which is less than optimal UX. Having just
one select box containing both month and year for the allowed
dates is better. This commit adds that, along with some small
tidy ups to the code. Closes #1235